+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './AesManager';
 export * from './CryptographyService';
 export * from './hmac';
+export * from './i18n/clinical-sections.i18n';
 export * from './storage';
 export * from './utils';