+++ src/models/interoperable-claims/index.ts
@@ -6,5 +6,8 @@ export * from './composition-claims';
 export * from './communication-claims';
 export * from './document-reference-claims';
 export * from './medication-statement-claims';
+export * from './allergy-intolerance-claims';
+export * from './condition-claims';
+export * from './device-use-statement-claims';
 export * from './appointment-claims';
 export * from './appointment-response-claims';